backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
python-multipart>=0.0.9
pytest>=8.0.0
httpx>=0.27.0
openai>=1.30.0
google-genai>=0.8.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
email-validator>=2.0.0
alembic>=1.13.0
psycopg[binary]>=3.1.0
